requirements.txt
fastapi
uvicorn
anthropic
twilio
python-dotenv
requests
